requirements.txt
argparse
Pillow